requirements.txt
websocket-client,
pytorch
